+++ contracts/mocks/token/ERC20/MockERC20.sol
@@ -1,16 +1,16 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.19;
 
-import '@openzeppelin/contracts/token/ERC20/ERC20.sol';
+import '../../../token/openzeppelin/ERC20/extensions/ERC20_165.sol';
 
 /**
- * @title MockERC20
+ * @title MockERC20_165
  * @author @NFTCulture
  */
-contract MockERC20 is ERC20 {
-    uint256 private immutable TOTAL_SUPPLY;
+contract MockERC20_165 is ERC20_165 {
+    uint256 private immutable _maxSupply;
 
-    constructor() ERC20('MockERC20', 'M20') {
-        TOTAL_SUPPLY = 1000000000;
+    constructor() ERC20('MockERC20_165', 'M20') {
+        _maxSupply = 1000000000;
     }
 }