+++ CHANGELOG.md
@@ -7,6 +7,9 @@ As of 12/13/2022, this repo has been renamed from "nftc-open-contracts" to "nftc
 ## Version -- 1.5.next [Not published]
 - TODO
 
+## Version -- 1.5.20
+- Implement ERC165 compliant version of ERC20.
+- Update the NFTSpecChecker and tests.
 
 ## Version -- 1.5.19
 - package lock cleanup